[1.5] 修正矩阵快速幂模版溢出问题

+++ src/Tools/MatrixFastPower.java
@@ -11,7 +11,7 @@ public class MatrixFastPower {
         for (int i = 0; i < m; i++)
             for (int j = 0; j < n; j++)
                 for (int k = 0; k < t; k++)
-                    c[i][j] = (c[i][j] + a[i][k] * b[k][j]) % MOD;
+                    c[i][j] = (int) ((c[i][j] + (long) a[i][k] * b[k][j]) % MOD);
         return c;
     }
